refactor: delete unused code and move file
