+++ packages/arb-token-bridge-ui/src/util/teleports/helpers.ts
@@ -10,7 +10,7 @@ import { MergedTransaction } from '../../state/app/state'
 import { FetchEthTeleportsFromSubgraphResult } from './fetchEthTeleportsFromSubgraph'
 import { TeleportFromSubgraph } from './fetchTeleports'
 import { AssetType } from '../../hooks/arbTokenBridge.types'
-import { Transaction } from '../../hooks/useTransactions'
+import { Transaction } from '../../types/Transactions'
 import { transformDeposit } from '../../state/app/utils'
 import { updateAdditionalDepositData } from '../deposits/helpers'
 import { fetchErc20Data } from '../TokenUtils'
